Move article page to new view directory
