+++ src/components/index.ts
@@ -38,4 +38,3 @@ export { default as AddressSummary } from "./AddressSummary";
 export { default as CheckoutPayment } from "./CheckoutPayment";
 export { default as ShippingAddressForm } from "./ShippingAddressForm";
 export { default as CheckoutReview } from "./CheckoutReview";
-export { default as ArticlePage } from "./ArticlePage";
