+++ src/views/Article/scss/index.scss
@@ -1,4 +1,4 @@
-@import "../../App/scss/variables.scss";
+@import "../../../components/App/scss/variables.scss";
 
 .article-page {
   &__container {
